refactor: move 'nearbyOffers' selectors to own folder
